Display CreateUsernameActivity instead of InviteFriendActivity after using invitation
