+++ wallet/res/layout/fragment_invitation_created.xml
@@ -78,7 +78,7 @@
 
                     <Button
                         android:id="@+id/preview_button"
-                        style="@style/DashButton.Transparent"
+                        style="@style/DashButton.Transparent.BlueText"
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
                         android:layout_gravity="center"
@@ -120,7 +120,7 @@
 
                 <Button
                     android:id="@+id/copy_invitation_link"
-                    style="@style/DashButton.Transparent"
+                    style="@style/DashButton.Transparent.BlueText"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_gravity="center"
