--- wallet/src/de/schildbach/wallet/ui/invite/InviteFriendViewModel.kt
@@ -1,32 +1,0 @@
-/*
- * Copyright 2021 Dash Core Group
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package de.schildbach.wallet.ui.invite
-
-import android.app.Application
-import androidx.lifecycle.AndroidViewModel
-import androidx.lifecycle.LiveData
-import de.schildbach.wallet.AppDatabase
-import de.schildbach.wallet.WalletApplication
-import de.schildbach.wallet.data.DashPayProfile
-
-open class InviteFriendViewModel(application: Application) : AndroidViewModel(application) {
-
-    val walletApplication = application as WalletApplication
-
-    fun dashPayProfileData(username: String): LiveData<DashPayProfile?> {
-        return AppDatabase.getAppDatabase().dashPayProfileDaoAsync().loadByUsernameDistinct(username)
-    }
-}
